Clean up mapped-objects
Some cases are not used any more.
Move logic into <related-comments> component.
